+++ test/test_nmt_master.py
@@ -151,15 +151,15 @@ class TestWorkspaceHPX(unittest.TestCase):
         self.lmax = 3*self.nside-1
         self.npix = int(hp.nside2npix(self.nside))
         self.msk = hp.read_map("test/benchmarks/msk.fits",
-                               verbose=False)
+                               verbose=False, dtype=float)
         self.mps = np.array(hp.read_map("test/benchmarks/mps.fits",
-                                        verbose=False,
+                                        verbose=False, dtype=float,
                                         field=[0, 1, 2]))
         self.mps_s1 = np.array(hp.read_map("test/benchmarks/mps_sp1.fits",
-                                           verbose=False,
+                                           verbose=False, dtype=float,
                                            field=[0, 1, 2]))
         self.tmp = np.array(hp.read_map("test/benchmarks/tmp.fits",
-                                        verbose=False,
+                                        verbose=False, dtype=float,
                                         field=[0, 1, 2]))
         self.b = nmt.NmtBin.from_nside_linear(self.nside, self.nlb)
         self.f0 = nmt.NmtField(self.msk,
